+++ src/WarpBusiness.Web/Program.cs
@@ -1,4 +1,5 @@
 using WarpBusiness.Web.Components;
+using WarpBusiness.Web.Services;
 
 var builder = WebApplication.CreateBuilder(args);
 
@@ -6,6 +7,12 @@ builder.AddServiceDefaults();
 
 builder.Services.AddRazorComponents()
     .AddInteractiveServerComponents();
+
+builder.Services.AddSingleton<AuthStateService>();
+builder.Services.AddHttpClient<WarpApiClient>(client =>
+{
+    client.BaseAddress = new Uri("https+http://api");
+});
 
 var app = builder.Build();